% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_names.R
\name{change_names}
\alias{change_names}
\title{Change names of columns}
\usage{
change_names(data, labelset, from, to)
}
\arguments{
\item{data}{the data frame that you want to change the name of}

\item{labelset}{the data frame containing labels that you want to change and what you want them to be changed to}

\item{from}{the name of the column in the label set data frame containing the original names}

\item{to}{the name of the column in the label set data frame containing new names that the original names will be changed to}
}
\value{
a data frame containing new column names
}
\description{
Using a new data frame of labels, change column names in one function
}
\examples{
#creating data set
Sites <- c("One", "Two", "Three", "Four", "Five")
Acrop <- c(0.1, 0.4, 0.9, 0.2, 0.5)
Gardin <- c(0.4, 0.9, 0.5, 0.23, 0.8)
Psam <- c(0.9, 0.5, 0.8, 0.1, 0.4)
Lepta <- c(0.5, 0.7, 0.8, 0.2, 0.9)
coral_cover <- data.frame(Sites, Acrop, Gardin, Psam, Lepta)

#creating label data frame
species_short <- c("Acrop", "Gardin", "Psam", "Lepta")
species_long <- c("Acropora", "Gardineroseris", "Psammocora", "Leptastrea")
coral_labels <- data.frame(species_short, species_long)

change_names(coral_cover, coral_labels, "species_short", "species_long")

}
