% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coral_labelset.R
\docType{data}
\name{coral_labelset}
\alias{coral_labelset}
\title{Labelset for Corals}
\format{
A \code{data_frame} with 5 columns, which are:
\describe{
\item{short_name}{The short name ID used to annotate the photo quadrats.}
\item{full_name}{The long name and definition for each short name ID.}
\item{taxonomic_name}{The taxonomic grouping name for each short name ID.}
\item{functional_group}{The functional group for each ID.}
\item{life_history}{The life history category for each coral ID.}
}
}
\usage{
coral_labelset
}
\description{
A \code{data_frame} that can be used with the Simple Cleaning Quadrat
Data vignette to rename columns in quadrat data and add functional group data,
to produce easy to analyze data frames. The short names were used to
identify substrate using \href{https://coralnet.ucsd.edu/}{CoralNet} and are
included in the output from CoralNet.
}
\keyword{datasets}
