% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadcleanR-package.R
\docType{package}
\name{quadcleanR-package}
\alias{quadcleanR}
\alias{quadcleanR-package}
\title{quadcleanR: Cleanup and Visualization of Quadrat Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool that can be customized to aid in the clean up of ecological data collected using quadrats and can crop quadrats to ensure comparability between quadrats collected under different methodologies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DominiqueMaucieri/quadcleanR}
  \item \url{https://dominiquemaucieri.com/quadcleanR/}
  \item Report bugs at \url{https://github.com/DominiqueMaucieri/quadcleanR/issues}
}

}
\author{
\strong{Maintainer}: Dominique Maucieri \email{dominiquemaucieri@gmail.com} (\href{https://orcid.org/0000-0003-1849-2472}{ORCID})

}
\keyword{internal}
